# Website/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduAI Platform API",
        "description": "API Documentation for EduAI Platform",
        "contact": {
            "email": "support@eduai.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/v1/quiz": {
            "get": {
                "tags": [
                    "5. Quiz"
                ],
                "summary": "List my quizzes",
                "description": "Returns all quizzes created by the authenticated user.",
                "operationId": "233de901c5d0a896149b060efb89de74",
                "responses": {
                    "200": {
                        "description": "List of quizzes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Chapter 3 Quiz"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "pending",
                                                            "processing",
                                                            "done",
                                                            "failed"
                                                        ],
                                                        "example": "done"
                                                    },
                                                    "is_public": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "share_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/quiz/public/uuid"
                                                    },
                                                    "questions_count": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "attempts_count": {
                                                        "type": "integer",
                                                        "example": 12
                                                    },
                                                    "source_type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "text",
                                                            "pdf",
                                                            "docx",
                                                            "pptx"
                                                        ],
                                                        "example": "pdf"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "5. Quiz"
                ],
                "summary": "Create a quiz",
                "description": "Create a quiz from raw text or upload a PDF/DOCX/PPTX file. Processing is asynchronous — poll the status endpoint.",
                "operationId": "deceef8d8d4aacca555ddc27696dd688",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Chapter 3 Quiz"
                                    },
                                    "max_questions": {
                                        "type": "integer",
                                        "maximum": 20,
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "text": {
                                        "description": "Raw text (required if no file)",
                                        "type": "string",
                                        "example": "Binary search is an efficient algorithm..."
                                    },
                                    "file": {
                                        "description": "Upload PDF/DOCX/PPTX (required if no text)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Quiz creation started",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Quiz creation started. Poll the status endpoint."
                                        },
                                        "quiz_id": {
                                            "type": "integer",
                                            "example": 7
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/quiz/{id}": {
            "get": {
                "tags": [
                    "5. Quiz"
                ],
                "summary": "Get quiz details",
                "description": "Returns full quiz details with all questions and correct answers. Owner only.",
                "operationId": "7f441290249d737b197daf4d79dcf862",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quiz details with questions"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Quiz not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/quiz/{id}/status": {
            "get": {
                "tags": [
                    "5. Quiz"
                ],
                "summary": "Poll quiz generation status",
                "description": "Poll this endpoint every 5 seconds after creating a quiz until status becomes 'done' or 'failed'.",
                "operationId": "93cdd37727d75c1b80e867cd753bd36c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Current status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "quiz_id": {
                                            "type": "integer",
                                            "example": 7
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "pending",
                                                "processing",
                                                "done",
                                                "failed"
                                            ],
                                            "example": "done"
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "questions_count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/quiz/{id}/toggle-share": {
            "post": {
                "tags": [
                    "5. Quiz"
                ],
                "summary": "Toggle public share link",
                "description": "Enable or disable the public shareable link for this quiz.",
                "operationId": "d0e31b016fe0e79162c25c9c993e495f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Share status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_public": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "share_url": {
                                            "type": "string",
                                            "example": "https://example.com/quiz/public/uuid"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Quiz is now public."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Session Cookie Authentication",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}